process: rename filesource to file_source
